Fix bug with Emacs
Emacs gets confused if a note begins with some colons because it thinks
that is specifying the line/column number, so then it parses the output
of the spaced inbox in such a way that the actual line number is part of
the file. To be more concrete, an output line like:

    C:\Users\Issa\inbox.txt:500:1:3:35pm:blah blah blah

*Should* be parsed as:

    filepath:      C:\Users\Issa\inbox.txt
    line number:   500
    column number: 1
    note fragment: 3:35pm:blah blah blah

But in fact gets parsed by Emacs as:

    filepath:      C:\Users\Issa\inbox.txt:500
    line number:   1
    column number: 3
    note fragment: 35pm:blah blah blah

My (possibly temporary) fix is to just remove all colons from the note
fragment, so that there is no ambiguity about what the line number is.